+++ src/dashboardWebView/components/SnippetsView/SnippetForm.tsx
@@ -79,15 +79,21 @@ const SnippetForm: React.ForwardRefRenderFunction<SnippetFormHandle, ISnippetFor
     const allFields: SnippetField[] = [];
     const snippetFields = snippet.fields || [];
 
-    for (const fieldName of placeholders) {
-      const field = snippetFields.find(f => f.name === fieldName);
-
-      if (field) {
+    // Loop over all fields to check if they are present in the snippet
+    for (const field of snippetFields) {
+      const idx = placeholders.findIndex(fieldName => fieldName === field.name);
+      if (idx > -1) {
         allFields.push({
           ...field,
           value: insertPlaceholderValues(field.default || "")
         });
-      } else {
+      }
+    }
+
+    // Loop over all placeholders to find the ones that are not present in the snippet fields
+    for (const fieldName of placeholders) {
+      const idx = snippetFields.findIndex(field => field.name === fieldName);
+      if (idx === -1) {
         allFields.push({
           name: fieldName,
           title: fieldName,